+++ src/main/java/br/com/dogsteps/repositories/TutorRepository.java
@@ -54,7 +54,7 @@ public class TutorRepository implements IRepository<Tutor, String>
         try {
             return TUTOR_DAO.update(tutor) ?
                     Response.status(Response.Status.OK).build() :
-                    Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+                    Response.status(Response.Status.NOT_FOUND).build();
         }catch (ValorNegativoException | StringVaziaException |
                     EmailInvalidoException | NullPointerException e){
             return Response.status(Response.Status.BAD_REQUEST).build();
@@ -62,40 +62,32 @@ public class TutorRepository implements IRepository<Tutor, String>
     }
     @Override
     public Response remove(String id){
-        if ( ! (id.isEmpty())) {
-            return TUTOR_DAO.remove(id) ?
-                    Response.status(Response.Status.OK).build() :
-                    Response.status(Response.Status.NOT_FOUND).build();
-        } return
+        if(id != null) {
+            if (!(id.isEmpty())) {
+                return TUTOR_DAO.remove(id) ?
+                        Response.status(Response.Status.OK).build() :
+                        Response.status(Response.Status.NOT_FOUND).build();
+            }
+        }return
                 Response.status(Response.Status.BAD_REQUEST).build();
     }
 
     private void validarRequisicao(Tutor tutor) throws ValorNegativoException, StringVaziaException,
                     EmailInvalidoException{
 
-        if(tutor.getIdade() > 0){ }
-        else{
+        if(tutor.getIdade() <= 0)
             throw new ValorNegativoException();
-        }
 
-        if( ! (tutor.getNome().isEmpty() || tutor.getEmail().isEmpty() || tutor.getCpf().isEmpty()
-                || tutor.getSenha().isEmpty() )
-        ) {}
-        else{
+        if( tutor.getNome().isEmpty() || tutor.getEmail().isEmpty()
+                || tutor.getCpf().isEmpty() || tutor.getSenha().isEmpty() )
             throw new StringVaziaException();
-        }
 
         String regex = "^[\\w-_\\.+]*[\\w-_\\.]\\@([\\w]+\\.)+[\\w]+[\\w]$";
-        if(tutor.getEmail().matches(regex)){ }
-        else{
+        if(!( tutor.getEmail().matches(regex) ))
             throw new EmailInvalidoException();
-        }
 
-        if(! (tutor.getAgenda() == null || tutor.getPasseios()  == null ||
-                tutor.getEndereco() == null) ){ }
-        else{
+        if(tutor.getAgenda() == null || tutor.getPasseios()  == null ||
+                tutor.getEndereco() == null)
             throw new NullPointerException();
-        }
-
     }
 }
